move PlayerSelctionMenuItems to one file and reset the states to default signature
