+++ Frequency/Menu/PlayerMenu/PlayerSelection.py
@@ -1,25 +1,14 @@
-﻿import pygame
-import sys
-from functools import reduce
-from pygame.surface import Surface
+﻿from functools import reduce
 
 import Game
-from Menu.PlayerMenu.PlayerMenuItems.TwoPlayers import TwoPlayers
-from Menu.PlayerMenu.PlayerMenuItems.ThreePlayers import ThreePlayers
-from Menu.PlayerMenu.PlayerMenuItems.FourPlayers import FourPlayers
-from Menu.StartMenu.StartMenuItems.StartMenuItem import StartMenuItem
+from Menu.HeadMenu import HeadMenu
+from Menu.PlayerMenu.PlayerSelectionMenuItems import *
 from Vector2 import Vector2
 
-class PlayerSelection(StartMenuItem):
+class PlayerSelection(HeadMenu):
 
-    def __init__(self, game, background=None, logo=None, playerMenuItems=None):
-        self.Resolution = game.Settings.Resolution
-
-        self.Background = background if background is not None \
-            else pygame.transform.scale(pygame.image.load('images/gameBackground.jpg'), [self.Resolution.X, self.Resolution.Y])
-
-        self.Logo = logo if logo is not None \
-            else pygame.transform.scale(pygame.image.load('images/gameLogo.png'), (230, 230))
+    def __init__(self, resolution, background=None, logo=None, playerMenuItems=None):
+        super().__init__(resolution, background, logo)
 
         self.PlayerMenuItems = playerMenuItems if playerMenuItems is not None \
             else [TwoPlayers(Vector2(0 ,0)), ThreePlayers(Vector2(0, 70)), FourPlayers(Vector2(0, 140))]
@@ -33,22 +22,7 @@ class PlayerSelection(StartMenuItem):
             else PlayerSelection(game, self.Background, self.Logo, newPlayerMenuItems)
 
     def Draw(self, game):
-        # Extra screen-based properties
-        screen_centerX = int(game.Settings.Resolution.X // 2)
-        screen_marginX = int(game.Settings.Resolution.Y / 18)
-       
-        # Logo position
-        logoSize = self.Logo.get_rect()
-        logoCenterX = screen_centerX - logoSize.centerx
-        logoCenter = (logoCenterX, screen_marginX)
-
-         # Basic screen loaded elements
-        game.Settings.GetScreen().fill((255, 255, 255))
-        game.Settings.GetScreen().blit(self.Background, (0, 0))
-        game.Settings.GetScreen().blit(self.Logo, logoCenter)
+        super().Draw(game)
 
         for menuItem in self.PlayerMenuItems:
             menuItem.Draw(game)
-
-    def GetNewState(self):
-        return None # will be the new state
